DL-7631 - removed legacy UI components
