+++ app/views/calculation/annualExemptAmount.scala.html
@@ -23,7 +23,7 @@
     layout: Layout,
     formInputMoney: playComponents.formInputMoney,
     errorSummary: playComponents.errorSummary,
-        submitButton: playComponents.submitButton
+    submitButton: playComponents.submitButton
 )
 
 @(annualExemptAmountForm: Form[AnnualExemptAmountModel], maxAEA: BigDecimal, backUrl: String)(implicit request: Request[_], messages: Messages)
