+++ app/views/calculation/improvements.scala.html
@@ -18,7 +18,6 @@
 @import uk.gov.hmrc.govukfrontend.views.html.components.FormWithCSRF
 
 @this(
-    mainTemplate: main_template,
     layout: Layout,
     errorSummary: playComponents.errorSummary,
     form: FormWithCSRF,
