+++ app/views/calculation/privateResidenceRelief.scala.html
@@ -19,18 +19,14 @@
 @import common.Dates
 @import common.Dates.TemplateImplicits._
 @import uk.gov.hmrc.govukfrontend.views.html.components.FormWithCSRF
-@import uk.gov.hmrc.play.views.html.helpers.Input
 
 @this(
-    uiInput: Input,
-    formHiddenYesNoRadio: helpers.formHiddenYesNoRadio,
     formHiddenYesNoRadioCurrencyInput : playComponents.formHiddenYesNoRadioCurrencyInput,
     inputYesNo : playComponents.input_yes_no,
     form: FormWithCSRF,
     layout: Layout,
     errorSummary: playComponents.errorSummary,
-    submitButton: playComponents.submitButton,
-    expandableHelpTextHelper: helpers.expandableHelpTextHelper
+    submitButton: playComponents.submitButton
 )
 
 @(privateResidenceReliefForm: Form[PrivateResidenceReliefModel], daysBetweenShow: Boolean, showFirstQuestion: Boolean, disposalDateLess18Months: Option[LocalDate], pRRMonths: Int, showOnlyFlatQuestion: Boolean)(implicit request: Request[_], messages: Messages, lang: Lang)
