--- app/views/helpers/calculationElectionHelperForm.scala.html
@@ -1,94 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import scala.concurrent.ExecutionContext
-@import helper._
-@import uk.gov.hmrc.play.views.helpers.MoneyPounds
-@import play.api.i18n._
-@import scala.concurrent.Future
-@import uk.gov.hmrc.play.views.formatting.Money._
-
-@this(implicit ex: ExecutionContext)
-
-@(field: Field, radioOptions: Seq[(String, String, String, String, Option[String], Option[BigDecimal])], showErrorText: Boolean, args: (Symbol, Any)*)(implicit messages: Messages)
-
-@elements = @{new FieldElements(field.id, field, null, args.toMap, messages) }
-@fieldsetClass = {@elements.args.get('_groupClass)}
-@labelAfter = @{ elements.args.get('_labelAfter).getOrElse(true).asInstanceOf[Boolean] }
-<div class="@if(elements.hasErrors){ form-field--error}">
-    <fieldset id="@field.id" class="@fieldsetClass"
-        @if(elements.args.get('_fieldsetAttributes).isDefined) {@elements.args.get('_fieldsetAttributes)}>
-        @if(elements.args.get('_legend).isDefined) {
-            <legend @if(elements.args.get('_legendClass).isDefined) {class="@elements.args.get('_legendClass)"}
-                    @if(elements.args.get('_legendID).isDefined) {id="@elements.args.get('_legendID)"}>
-                @elements.args.get('_legend)
-            </legend>
-        }
-        @if(elements.args.get('_helpText).isDefined) {
-            <span class="form-hint">
-                @elements.args.get('_helpText)
-            </span>
-        }
-        @elements.errors.map{error => <span class="error-notification">@if(showErrorText){ @Messages("calc.base.optionReqError")}</span>}
-
-        @radioOptions.map { case (value, amount, message, calcType, dateMessage, otherReliefs) =>
-            @defining(s"${elements.field.name}-${value.toLowerCase.replace(" ","_")}")  { inputId =>
-               <div>
-                <label for="@inputId"
-                    @elements.args.get('_labelClass).map{labelClass => class="@labelClass@field.value.filter( _ == value).map{_ => selected}"}>
-                    <input
-                        type="radio"
-                        id="@inputId"
-                        name="@elements.field.name"
-                        value="@value"
-                        aria-expanded="false"
-                        @elements.args.get('_inputClass).map{inputClass => class="@inputClass"}
-                        @if(elements.args.contains('_dataAttributes) ){ @elements.args.get('_dataAttributes)}
-                        @field.value.filter( _ == value).map{_ => checked="checked"} />
-                    <span class="bold-small">@Messages(calcType)</span>
-                    @defining(s"${value.toLowerCase.replace(" ","_")}-para") { radioParaId =>
-                            <div class="form-group">
-                                <span class="election-text" id="@radioParaId">
-                                    <span>@Messages(message)</span>
-                                    @if(dateMessage.isDefined){<span class="no-wrap">@Messages(dateMessage.get)</span>}
-                                </span>
-                            </div>
-                            <div>
-                                <span class="bold-small">&pound;@MoneyPounds(BigDecimal(amount)).quantity</span>
-                            </div>
-                    }
-                </label>
-
-                @{
-                    if(otherReliefs.exists(_ > 0)){
-                        <div class="panel panel-indent additional-option-block">
-                            <span class="align-bottom">{Messages("calc.calculationElection.someOtherRelief")}</span>
-                            <button id={s"$value-button"} name="action" value={value} class="button button--link">
-                                <span class="bold-medium">&pound;{MoneyPounds(otherReliefs.getOrElse(0)).quantity}</span>
-                            </button>
-                        </div>
-                        } else if (otherReliefs.isDefined) {
-                        <div class="panel panel-indent additional-option-block">
-                            <button id={s"$value-button"} name="action" value={value} class="button button--link">{Messages("calc.calculationElection.otherRelief")}</button>
-                        </div>
-                    }
-                }
-                </div>
-            }
-        }
-
-    </fieldset>
-</div>
