--- app/views/helpers/errorSummary.scala.html
@@ -1,57 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@this()
-
-@(form: Form[_], page: String, errorKey: String = "", additionalText: Option[String] = None)(implicit messages: Messages)
-
-@summary = {
-    <div
-    class="flash error-summary error-summary--show"
-    id="error-summary-display"
-    role="alert"
-    aria-labelledby="error-summary-heading"
-    tabindex="-1">
-
-        <h2 id="error-summary-heading" class="h3-heading">@Messages("calc.error.summary.heading")</h2>
-        @if(additionalText.isDefined) {
-            <p>@additionalText</p>
-        }
-        <ul class="js-error-summary-messages">
-        @form.errors.map { error =>
-            @if(error.key.equals("")) {
-                <li role="tooltip" data-journey="@page:error:@errorKey">
-                    <a href="#@errorKey"
-                    id="@{errorKey}-error-summary"
-                    data-focuses="@errorKey">
-                    @Messages(error.message, error.args: _*)
-                    </a>
-                </li>
-            } else {
-                <li role="tooltip" data-journey="@page:error:@error.key">
-                    <a href="#@error.key"
-                    id="@{error.key}-error-summary"
-                    data-focuses="@error.key">
-                    @Messages(error.message, error.args: _*)
-                    </a>
-                </li>
-            }
-        }
-        </ul>
-    </div>
-}
-
-@(if(form.hasErrors) summary)
