--- app/views/helpers/formHiddenYesNoRadio.scala.html
@@ -1,63 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@this(
-    inputRadioGroup: InputRadioGroup
-)
-
-@(form: Form[_], fieldName: String, questionText:String, content:Html, helpText:Option[String] = None, hideLegend: Boolean = false, hidePanel: Boolean = false, additionalContent: Option[Html] = None, labelClass: Boolean = false)(implicit messages: Messages)
-
-@hiddenClasses = @{if(form.errors.filter(_.key == "").length > 0) "form-field--error" else "" }
-
-<div class="form-group" data-hidden='hidden'>
-
-    @if(hideLegend) {
-        @inputRadioGroup(
-            field = form(fieldName),
-            Seq(
-            "Yes"->Messages("calc.base.yes"),
-            "No"->Messages("calc.base.no")),
-            '_legend -> questionText,
-            '_legendClass -> "visuallyhidden",
-            '_legendID -> fieldName,
-            '_helpText -> helpText,
-            '_labelAfter -> true,
-            '_labelClass -> "block-label",
-            '_groupClass -> "inline"
-        )
-    } else {
-        @inputRadioGroup(
-            field = form(fieldName),
-            Seq(
-                "Yes"->Messages("calc.base.yes"),
-                "No"->Messages("calc.base.no")),
-            '_legend -> questionText,
-            '_legendID -> fieldName,
-            '_helpText -> helpText,
-            '_labelAfter -> true,
-            '_labelClass -> "block-label",
-            '_groupClass -> "inline"
-        )
-    }
-
-    <div class="@{if (!hidePanel) "panel-indent panel-indent--wide panel-indent--gutter"}" id='hidden'>
-        <div class="@hiddenClasses">
-            @form.errors.filter(_.key == "").map { error => <span class="error-notification">@Messages(error.message, error.args: _*)</span>}
-            @if(additionalContent.isDefined) { @additionalContent }
-            @content
-        </div>
-    </div>
-</div>
