--- app/views/helpers/formInlineDateInput.scala.html
@@ -1,78 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import uk.gov.hmrc.play.views.html.helpers.Input
-
-@this(
-    uiInput: Input
-)
-
-@(formItem:Form[_], fieldName: String, questionText: String, args: (Symbol,Any)*)(implicit messages: Messages)
-
-@import uk.gov.hmrc.play.mappers.DateFields._
-@import play.api.i18n._
-@import java.text.SimpleDateFormat
-
-@fieldsetClasses = @{if(formItem.hasErrors && (formItem.errors.find(_.key == "").isDefined || formItem.errors.find(_.key != "hasAcquisitionDate").isDefined)) "form-field--error" else "" }
-
-    <fieldset class="form-group form-date @fieldsetClasses" id="@fieldName-fieldset">
-
-    <legend @{if(args.toMap.contains('_visually_hidden)) {"class=visuallyhidden"}}>
-        @questionText
-    </legend>
-
-    @formItem.errors.map { error => <span class="error-notification">@Messages(error.message)</span>}
-
-    @if(args.toMap.contains('_additionalHint)) {
-        <span class ="form-hint">@{args.toMap.get('_additionalHint)}</span>
-    }
-    <span class="form-hint">@Messages("calc.common.date.hint")</span>
-
-
-    @uiInput(
-        formItem(s"${fieldName}Day"),
-        '_label -> Messages("calc.common.date.fields.day"),
-        '_labelClass -> "form-group form-group-day",
-        '_type -> "number",
-        '_inputClass -> s"input--xxsmall input--no-spinner",
-        '_maxlength -> "2",
-        '_hideErrors -> " ",
-        '_hideBar -> " "
-    )
-
-    @uiInput(
-        formItem(s"${fieldName}Month"),
-        '_label -> Messages("calc.common.date.fields.month"),
-        '_labelClass -> "form-group form-group-month",
-        '_type -> "number",
-        '_inputClass -> s"input--xxsmall input--no-spinner",
-        '_maxlength -> "2",
-        '_hideErrors -> " ",
-        '_hideBar -> " "
-    )
-
-    @uiInput(
-        formItem(s"${fieldName}Year"),
-        '_label -> Messages("calc.common.date.fields.year"),
-        '_labelClass -> "form-group form-group-year",
-        '_type -> "number",
-        '_inputClass -> s"input--small input--no-spinner",
-        '_maxlength -> "4",
-        '_hideErrors -> " ",
-        '_hideBar -> " "
-    )
-
-</fieldset>
