--- app/views/helpers/formInputMoney.scala.html
@@ -1,75 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import play.api.data._
-
-@this(
-    moneyInputFormat: helpers.moneyInputFormat
-)
-
-@(inputForm: Form[_], fieldName: String, questionText: String, helpText: Option[String] = None, additionalHTMLContent: Option[String] = None, hideLabel: Boolean = false, step: String = "0.01", labelClasses: String = "")(implicit messages: Messages, messagesApi: Option[MessagesApi] = None)
-
-    <div class="form-group">
-
-        @if(hideLabel) {
-            @if(additionalHTMLContent.isDefined) {
-                @moneyInputFormat(
-                    inputForm(fieldName),
-                    '_label -> questionText,
-                    '_labelTextClass -> "visuallyhidden",
-                    '_additionalContent -> Html(additionalHTMLContent.get),
-                    '_inputHint -> Html(helpText.getOrElse("")),
-                    '_type -> "number",
-                    '_inputClass -> "input--no-spinner",
-                    '_step -> step
-                )
-            } else {
-                @moneyInputFormat(
-                    inputForm(fieldName),
-                    '_label -> questionText,
-                    '_labelTextClass -> "visuallyhidden",
-                    '_inputHint -> Html(helpText.getOrElse("")),
-                    '_type -> "number",
-                    '_inputClass -> "input--no-spinner",
-                    '_step -> step
-                )
-            }
-        } else {
-            @if(additionalHTMLContent.isDefined) {
-                @moneyInputFormat(
-                    inputForm(fieldName),
-                    '_label -> questionText,
-                    '_labelTextClass -> labelClasses,
-                    '_additionalContent -> Html(additionalHTMLContent.get),
-                    '_inputHint -> Html(helpText.getOrElse("")),
-                    '_type -> "number",
-                    '_inputClass -> "input--no-spinner",
-                    '_step -> step
-                )
-            } else {
-                @moneyInputFormat(
-                    inputForm(fieldName),
-                    '_label -> questionText,
-                    '_labelTextClass -> labelClasses,
-                    '_inputHint -> Html(helpText.getOrElse("")),
-                    '_type -> "number",
-                    '_inputClass -> "input--no-spinner",
-                    '_step -> step
-                )
-            }
-        }
-
-    </div>
