--- app/views/helpers/formInputMoneyHTML.scala.html
@@ -1,48 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import play.api.data._
-
-@this(
-    moneyInputFormat: helpers.moneyInputFormat
-)
-
-@(inputForm: Form[_], fieldName: String, questionText: String, helpText: Option[Html] = None, hideLabel: Boolean = false, step: String = "0.01")(implicit messages: Messages)
-
-<div class="form-group">
-
-    @if(hideLabel) {
-        @moneyInputFormat(
-            inputForm(fieldName),
-            '_label -> questionText,
-            '_labelTextClass -> "visuallyhidden",
-            '_inputHint -> helpText,
-            '_type -> "number",
-            '_inputClass -> "input--no-spinner",
-            '_step -> step
-        )
-    } else {
-        @moneyInputFormat(
-            inputForm(fieldName),
-            '_label -> questionText,
-            '_inputHint -> helpText,
-            '_type -> "number",
-            '_inputClass -> "input--no-spinner",
-            '_step -> step
-        )
-    }
-
-</div>
