--- app/views/helpers/formYesNoRadio.scala.html
@@ -1,53 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import play.api.i18n._
-
-@this(
-    inputRadioGroup: InputRadioGroup
-)
-
-@(field: Field, legend: String, legendClass: Option[String] = None)(implicit messages: Messages, messagesApi: Option[MessagesApi] = None)
-
-@{legendClass match {
-    case Some(legendClass) => {
-        inputRadioGroup(
-            field = field,
-            Seq(
-            "Yes" -> Messages("calc.base.yes"),
-            "No" -> Messages("calc.base.no")),
-            '_legend -> legend,
-            '_legendID -> field.id,
-            '_legendClass -> legendClass,
-            '_labelAfter -> true,
-            '_labelClass -> "block-label",
-            '_groupClass -> "inline form-group radio-list",
-            '_fieldsetAttributes -> "aria-details = help")
-    }
-    case _ => {
-        inputRadioGroup(
-            field = field,
-            Seq(
-            "Yes" -> Messages("calc.base.yes"),
-            "No" -> Messages("calc.base.no")),
-            '_legend -> legend,
-            '_legendID -> field.id,
-            '_labelAfter -> true,
-            '_labelClass -> "block-label",
-            '_groupClass -> "inline form-group radio-list",
-            '_fieldsetAttributes -> "aria-details = help")
-    }
-}}
