--- app/views/helpers/full_width_banner.scala.html
@@ -1,28 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@this()
-
-@(fullWidthBannerContent: Option[Html])(implicit messages: Messages)
-
-<div id="ur-panel" class="banner-panel">
-    <section class="centered-content visible banner-panel__container" aria-hidden="false">
-        @fullWidthBannerContent
-        <a class="banner-panel__close" href="#" role="button"
-           data-journey-click="link - click:UR Banner:Dismiss"
-        ><span aria-hidden="true">@Messages("calc.summary.dismissBanner")</span><span class="visually-hidden">@Messages("calc.summary.dismissBanner.screenreader")</span></a>
-    </section>
-</div>
