--- app/views/helpers/hiddenHelpText.scala.html
@@ -1,28 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@this()
-
-@(helpTextId: String, helpTextTitle: String, content: Html, additionalClasses: String = "")
-
-<div id="@helpTextId" class="form-group @additionalClasses">
-    <details>
-        <summary><span class="summary" data-journey-click="help:reveal:@helpTextTitle">@helpTextTitle</span></summary>
-        <div class="panel-indent panel-indent--wide panel-indent--gutter">
-            @content
-        </div>
-    </details>
-</div>
