--- app/views/helpers/moneyInputFormat.scala.html
@@ -1,104 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import helper._
-
-@this()
-
-@(field: play.api.data.Field, args: (Symbol,Any)*)(implicit messages: Messages)
-
-@import play.api.i18n._
-
-@elements = @{ new FieldElements(field.id, field, null, args.toMap, messages) }
-@parentField = @{args.toMap.get('parentField).asInstanceOf[Option[Field]]}
-@parentElements = @{parentField.map(pf => new FieldElements(pf.id, pf, null, Map(), messages) )}
-@value = @{ field.value match { case Some(x) => x case None => "" case x => x }}
-@labelAfter = @{ elements.args.get('_labelAfter).getOrElse(false).asInstanceOf[Boolean] }
-@labelHighlight = @{ elements.args.get('_labelHighlight).getOrElse(false).asInstanceOf[Boolean] }
-@additionalHTMLContent = @{ elements.args.contains('_additionalContent) }
-
-
-<label for="@elements.field.name" class="@elements.args.get('_divClass) @if( elements.args.get('_labelClass) ){ @elements.args.get('_labelClass) } @if(elements.hasErrors || (parentElements.isDefined && parentElements.get.hasErrors)) {form-field--error}" @if(elements.args.contains('_labelDataAttributes)){@elements.args.get('_labelDataAttributes)}>
-
-    @if(!elements.args.contains('_hideErrors)) {
-        @elements.errors.map { error =>
-            <span class="error-notification"
-                @if(elements.args.contains('_error_id)) {
-                    id="@elements.args.get('_error_id)"
-                }
-                role="tooltip">
-                @Messages(error.toString())
-            </span>
-        }
-    }
-
-    @if(parentElements.isDefined) {
-        @parentElements.get.errors.map { error => <span class="error-notification">@Messages(error.toString())</span>}
-    }
-
-    @if(!labelAfter && elements.args.contains('_label)) {
-        @if(labelHighlight){<strong>}
-
-        @if(elements.args.contains('_labelTextClass)) {
-            <div class="@elements.args.get('_labelTextClass)">@Messages(elements.label.toString)</div>
-        } else {
-            <div class="money-label">@Messages(elements.label.toString)</div>
-        }
-
-        @if(additionalHTMLContent) {
-            <div class="indent">@elements.args.get('_additionalContent)</div>
-        }
-
-        @if(elements.args.getOrElse('_inputHint, "").toString() != ""){
-            <span class="form-hint"
-                @if(elements.args.contains('_hintId)) {
-                    id="@elements.args.get('_hintId)"}>
-                @elements.args.get('_inputHint)
-            </span>
-        }
-        @if(labelHighlight){</strong>}
-    }
-
-    <span class="poundSign">&pound;</span>
-    <input
-        @if( elements.args.contains('_type)){type="@elements.args.get('_type)" }else{type="text" }
-        @if( elements.args.contains('_dataAttributes) ){ @elements.args.get('_dataAttributes)}
-        class="moneyField @if( elements.args.get('_inputClass) ){ @elements.args.get('_inputClass) }"
-        @if( elements.args.get('_autoComplete) ){ autocomplete="@elements.args.get('_autoComplete)" }
-        name="@elements.field.name"
-        id="@elements.field.name"
-        value="@value"
-        @if(elements.args.get('_error_id).isDefined) { aria-labeledby="@elements.args.get('_error_id)" }
-        @if(elements.args.get('_hintId).isDefined) { aria-describedby="@elements.args.get('_hintId)" }
-        @if(elements.args.get('_maxlength).isDefined) { maxlength="@elements.args.get('_maxlength)" }
-        @if(elements.args.get('_pattern).isDefined) { pattern="@elements.args.get('_pattern)" }
-        @if(elements.args.get('_title).isDefined) { title="@elements.args.get('_title)" }
-        @if(elements.args.get('_step).isDefined) { step="@elements.args.get('_step)" }
-        @if(elements.args.get('_required).isDefined) { required }
-    />
-
-    @if(labelAfter && elements.args.contains('_label)) {
-        @if(labelHighlight){<strong>}
-        <span @if(elements.args.contains('_labelTextClass)) { class="@elements.args.get('_labelTextClass)"}>
-            @Messages(elements.label.toString)
-        </span>
-        @if(elements.args.contains('_inputHint) ){
-            <span class="form-hint">@elements.args.get('_inputHint)</span>
-        }
-        @if(labelHighlight){</strong>}
-    }
-
-</label>
