--- app/views/helpers/summaryNumericRow.scala.html
@@ -1,86 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import uk.gov.hmrc.play.views.helpers.MoneyPounds
-
-@this()
-
-@(rowId: String, question: String, amount: BigDecimal, isTotal: Boolean = false, hintText: Boolean = true)(implicit messages: Messages, messagesApi: Option[MessagesApi] = None)
-
-@(isTotal, hintText) match {
-    case (true, _) => {
-
-        <div id="@rowId" class="grid-layout grid-layout--no-margin font-small summary-total-margin-top">
-            <div id="@{rowId}-text" class="grid-layout__column--1-2">
-                <strong>@question</strong>
-            </div>
-            <div id="@{rowId}-amount" class="grid-layout__column--1-3 text--right">
-            @amount.scale match {
-                case 0 => {
-                    <strong>&pound;@MoneyPounds(amount, 0).quantity</strong>
-                }
-                case _ if amount == amount.setScale(0, BigDecimal.RoundingMode.FLOOR) => {
-                    <strong>&pound;@MoneyPounds(amount, 0).quantity</strong>
-                }
-                case _ => {
-                    <strong>&pound;@MoneyPounds(amount, 2).quantity</strong>
-                }
-            }
-            </div>
-        </div>
-    }
-    case (_, true) => {
-        <div id="@rowId" class="grid-layout grid-layout--no-margin font-small hint-text">
-            <div id="@{rowId}-text" class="grid-layout__column--1-2">
-                @question
-            </div>
-            <div id="@{rowId}-amount" class="grid-layout__column--1-3 text--right">
-            @amount.scale match {
-                case 0 => {
-                    <span>&pound;@MoneyPounds(amount, 0).quantity</span>
-                }
-                case _ if amount == amount.setScale(0, BigDecimal.RoundingMode.FLOOR) => {
-                    <span>&pound;@MoneyPounds(amount, 0).quantity</span>
-                }
-                case _ => {
-                    <span>&pound;@MoneyPounds(amount, 2).quantity</span>
-                }
-            }
-            </div>
-        </div>
-    }
-    case (_, false) => {
-        <div id="@rowId" class="grid-layout grid-layout--no-margin font-small">
-            <div id="@{rowId}-text" class="grid-layout__column--1-2">
-            @question
-            </div>
-            <div id="@{rowId}-amount" class="grid-layout__column--1-3 text--right">
-            @amount.scale match {
-                case 0 => {
-                    <span>&pound;@MoneyPounds(amount, 0).quantity</span>
-                }
-                case _ if amount == amount.setScale(0, BigDecimal.RoundingMode.FLOOR) => {
-                    <span>&pound;@MoneyPounds(amount, 0).quantity</span>
-                }
-                case _ => {
-                    <span>&pound;@MoneyPounds(amount, 2).quantity</span>
-                }
-            }
-            </div>
-        </div>
-    }
-}
-
