--- app/views/helpers/summaryPartialDeductionsSection.scala.html
@@ -1,41 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@this(summaryNumericRow: summaryNumericRow)
-
-@(reliefsUsed: BigDecimal, aeaUsed: BigDecimal, inYearLossesUsed: BigDecimal, broughtForwardLossesUsed: BigDecimal,
-totalDeductions: BigDecimal)(implicit request: Request[_], messages: Messages, messagesApi: Option[MessagesApi] = None)
-
-<div id="deductionsSection" class="summary-underline summary-details-padding">
-
-    <h3>@Messages("calc.summaryPartial.deductionsSectionHeading")</h3>
-
-    @if(reliefsUsed > 0) {
-        @summaryNumericRow("reliefsUsed", Messages("calc.summaryPartial.reliefsUsed"), reliefsUsed)
-    }
-
-    @if(inYearLossesUsed > 0) {
-        @summaryNumericRow("inYearLossesUsed",Messages("calc.summaryPartial.inYearLossesUsed"), inYearLossesUsed)
-    }
-
-    @if(broughtForwardLossesUsed > 0) {
-        @summaryNumericRow("broughtForwardLossesUsed",Messages("calc.summaryPartial.broughtForwardLossesUsed"), broughtForwardLossesUsed)
-    }
-
-    @summaryNumericRow("aeaUsed",Messages("calc.summaryPartial.aeaUsed"), aeaUsed)
-
-    @summaryNumericRow("totalDeductions",Messages("calc.summaryPartial.totalDeductions"), totalDeductions, isTotal = true)
-</div>
