--- app/views/helpers/summarySectionNoLink.scala.html
@@ -1,36 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import models.QuestionAnswerModel
-
-@this(questionAnswerRowNoLink: questionAnswerRowNoLink)
-
-@(sectionID: String, sectionTitle: String, rows: Seq[QuestionAnswerModel[Any]], decimalPlaces: Int = 2)(implicit messages: Messages, lang: Lang)
-
-@if(rows.nonEmpty) {
-    <section id=@sectionID class="summary-section">
-    @for(row <- rows) {
-        <div class="grid-layout grid-layout--stacked form-group">
-            <div class="grid-layout__column grid-layout__column--1-3 pdf-header">
-            @if(row == rows.head) {
-                <span class="heading-large">@sectionTitle</span>
-            }
-            </div>
-            @questionAnswerRowNoLink(row, decimalPlaces)
-        </div>
-    }
-    </section>
-}
