--- app/views/main_template.scala.html
@@ -1,95 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import config.ApplicationConfig
-@import controllers.CgtLanguageController
-@import helpers.{full_width_banner, full_width_banner_content_ur}
-@import play.api.Application
-
-@this(
-    application: Application,
-    applicationConfig: ApplicationConfig,
-    uiSidebar: Sidebar,
-    uiArticle: Article,
-    languageSelection: helpers.language_selection,
-    fullWidthBanner: helpers.full_width_banner,
-    fullWidthBannerContentUR: helpers.full_width_banner_content_ur,
-    govukWrapper: govuk_wrapper
-)
-
-@(title: String,
-  sidebarLinks: Option[Html] = None,
-  contentHeader: Option[Html] = None,
-  bodyClasses: Option[String] = None,
-  mainClass: Option[String] = None,
-  scriptElem: Option[Html] = None,
-  isUserResearchBannerVisible: Boolean = false,
-  articleLayout: Boolean = true)(mainContent: Html)(implicit request : Request[_], messages: Messages)
-
-@serviceInfoContent = {}
-
-@lc = @{Application.instanceCache[CgtLanguageController].apply(application) }
-
-@sidebarContent = {
-    <div class="service-info">
-        <h2 class="heading-medium">@Messages("calc.common.readMore")</h2>
-        @sidebarLinks.get
-    </div>
-}
-
-@sidebar = {
-    @if(sidebarLinks.isDefined) {
-        @uiSidebar(sidebarContent, Some("sidebar"))
-    }
-}
-
-@contentLayout = {
-    @if(articleLayout) {
-        @uiArticle(mainContent)
-    } else {
-        @mainContent
-    }
-}
-
-@contentHeader = {
-    <div class="float--right">@languageSelection(lc.languageMap, controllers.routes.CgtLanguageController.switchToLanguage, None)(messages)</div>
-}
-
-@afterHeader = {
-    @if(isUserResearchBannerVisible) {
-
-    @fullWidthBanner(Some(fullWidthBannerContentUR(
-                messages("calc.summary.userResearchLink.title"),
-                messages("calc.summary.userResearchLink.link"),
-                applicationConfig.urBannerLink,
-                Some(messages("calc.summary.userResearchLink.text"))
-                )
-            )
-        )
-    }
-}
-
-@govukWrapper(title = title,
-               mainClass = mainClass,
-               bodyClasses = bodyClasses,
-               sidebar = sidebar,
-               contentHeader = Some(contentHeader),
-               mainContent = contentLayout,
-               serviceInfoContent = serviceInfoContent,
-               scriptElem = scriptElem,
-               afterHeader = afterHeader
-)
-
