--- app/views/playComponents/formMultipleInputMoney.scala.html
@@ -1,48 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import play.api.data._
-
-@this(
-    formInputMoney: helpers.formInputMoney
-)
-
-@(inputForm: Form[_], inputDetails: Seq[(String, String, Option[String])], boldText:Boolean = false)(implicit messages: Messages, messagesApi: Option[MessagesApi])
-
-<div class="form-group">
-
-    @inputDetails.map(index => {
-
-    if(!boldText) {
-        formInputMoney(
-            inputForm,
-            index._1, //fieldName
-            index._2, //questionText
-            index._3 //helpText
-        )
-    }
-        else {
-            formInputMoney(
-            inputForm,
-            index._1, //fieldName
-            index._2, //questionText
-            index._3,  //helpText
-            labelClasses = "bold-small"
-            )
-        }
-    })
-
-</div>
