+++ app/views/playComponents/questionAnswerRow.scala.html
@@ -1,18 +1,18 @@
 @*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
+* Copyright 2021 HM Revenue & Customs
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*     http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*@
 
 @import uk.gov.hmrc.play.views.helpers.MoneyPounds
 @import models.QuestionAnswerModel
@@ -20,23 +20,23 @@
 @import common.Dates._
 @import common.Dates.TemplateImplicits._
 
-@this(questionAnswerRowNoLink: helpers.questionAnswerRowNoLink)
+@this(questionAnswerRowNoLink: playComponents.questionAnswerRowNoLink)
 
 @(questionAnswer: QuestionAnswerModel[Any], decimalPlaces: Int)(implicit messages: Messages, lang: Lang)
 
 @questionAnswer.link.fold(questionAnswerRowNoLink(questionAnswer, decimalPlaces)) { link =>
-    <div id="@{s"${questionAnswer.id}-question"}" class="grid-layout__column grid-layout__column--1-2">
-        <span class="lede">@questionAnswer.question</span>
-    </div>
-    <div id="@questionAnswer.id-answer" class="grid-layout__column grid-layout__column--1-2">
-        <a class="lede summary-answer" href="@link">@{
-            questionAnswer.data match {
-                    case string: String => string
-                    case int: Int => int.toString
-                    case bigDecimal: BigDecimal => s"£${MoneyPounds(bigDecimal, decimalPlaces).quantity}"
-                    case date: LocalDate => date.localFormat("d MMMM yyyy")
-                    case _ => ""
-                }
-            }</a>
-    </div>
+<div id="@{s"${questionAnswer.id}-question"}" class="grid-layout__column grid-layout__column--1-2">
+<span class="lede">@questionAnswer.question</span>
+</div>
+<div id="@questionAnswer.id-answer" class="grid-layout__column grid-layout__column--1-2">
+    <a class="lede summary-answer" href="@link">@{
+        questionAnswer.data match {
+        case string: String => string
+        case int: Int => int.toString
+        case bigDecimal: BigDecimal => s"£${MoneyPounds(bigDecimal, decimalPlaces).quantity}"
+        case date: LocalDate => date.localFormat("d MMMM yyyy")
+        case _ => ""
+        }
+        }</a>
+</div>
 }
