+++ app/views/whatNext/whatNext.scala.html
@@ -14,7 +14,6 @@
  * limitations under the License.
  *@
 
-@import views.html.helpers.expandableHelpTextHelper
 @import config.ApplicationConfig
 
 @this(
