+++ test/views/playComponents/SummaryPartialTaxableGainViewSpec.scala
@@ -23,8 +23,7 @@ import org.jsoup.Jsoup
 import org.scalatestplus.mockito.MockitoSugar
 import play.api.i18n.Lang
 import play.api.mvc.MessagesControllerComponents
-import views.html.helpers
-import views.html.helpers.summaryPartialTaxableGain
+import views.html.playComponents.summaryPartialTaxableGain
 
 class SummaryPartialTaxableGainViewSpec extends CommonPlaySpec with WithCommonFakeApplication with FakeRequestHelper with MockitoSugar {
   implicit lazy val mockMessage = fakeApplication.injector.instanceOf[MessagesControllerComponents].messagesApi.preferred(fakeRequest)
