+++ test/views/playComponents/formYesNoRadioSpec.scala
@@ -23,7 +23,7 @@ import forms.BoughtForLessForm._
 import org.jsoup.Jsoup
 import org.scalatestplus.mockito.MockitoSugar
 import play.api.mvc.MessagesControllerComponents
-import views.html.helpers._
+import views.html.playComponents._
 
 class formYesNoRadioSpec extends CommonPlaySpec with WithCommonFakeApplication with MockitoSugar with FakeRequestHelper {
   implicit lazy val mockMessage = fakeApplication.injector.instanceOf[MessagesControllerComponents].messagesApi.preferred(fakeRequest)
